Seriously refactor catalog loading.
