--- src/lsdb/loaders/hats/association_catalog_loader.py
@@ -1,27 +1,0 @@
-import hats as hc
-import nested_dask as nd
-
-from lsdb.catalog.association_catalog import AssociationCatalog
-from lsdb.loaders.hats.abstract_catalog_loader import AbstractCatalogLoader
-
-
-class AssociationCatalogLoader(AbstractCatalogLoader[AssociationCatalog]):
-    """Loads an HATS AssociationCatalog"""
-
-    def load_catalog(self) -> AssociationCatalog:
-        """Load a catalog from the configuration specified when the loader was created
-
-        Returns:
-            Catalog object with data from the source given at loader initialization
-        """
-        hc_catalog = self._load_hats_catalog(hc.catalog.AssociationCatalog)
-        if hc_catalog.catalog_info.contains_leaf_files:
-            dask_df, dask_df_pixel_map = self._load_dask_df_and_map(hc_catalog)
-        else:
-            dask_df, dask_df_pixel_map = self._load_empty_dask_df_and_map(hc_catalog)
-        return AssociationCatalog(dask_df, dask_df_pixel_map, hc_catalog)
-
-    def _load_empty_dask_df_and_map(self, hc_catalog):
-        dask_meta_schema = self._create_dask_meta_schema(hc_catalog.schema)
-        ddf = nd.NestedFrame.from_pandas(dask_meta_schema, npartitions=1)
-        return ddf, {}
