+++ src/lsdb/loaders/hats/read_hats.py
@@ -1,26 +1,27 @@
 from __future__ import annotations
 
 from pathlib import Path
-from typing import Dict, List, Type
+from typing import List, Tuple
 
 import hats as hc
-from hats.catalog import CatalogType, TableProperties
+import nested_dask as nd
+import nested_pandas as npd
+import numpy as np
+import pyarrow as pa
+from hats.catalog import CatalogType
+from hats.catalog.healpix_dataset.healpix_dataset import HealpixDataset as HCHealpixDataset
+from hats.io.file_io import file_io
+from hats.pixel_math import HealpixPixel
+from hats.pixel_math.healpix_pixel_function import get_pixel_argsort
+from hats.pixel_math.spatial_index import SPATIAL_INDEX_COLUMN
 from upath import UPath
 
-from lsdb.catalog.margin_catalog import MarginCatalog
+from lsdb.catalog.association_catalog import AssociationCatalog
+from lsdb.catalog.catalog import Catalog, DaskDFPixelMap, MarginCatalog
 from lsdb.core.search.abstract_search import AbstractSearch
-from lsdb.loaders.hats.abstract_catalog_loader import AbstractCatalogLoader, CatalogTypeVar
-from lsdb.loaders.hats.association_catalog_loader import AssociationCatalogLoader
-from lsdb.loaders.hats.hats_catalog_loader import HatsCatalogLoader
+from lsdb.dask.divisions import get_pixels_divisions
 from lsdb.loaders.hats.hats_loading_config import HatsLoadingConfig
-from lsdb.loaders.hats.margin_catalog_loader import MarginCatalogLoader
-
-loader_class_for_catalog_type: Dict[CatalogType, Type[AbstractCatalogLoader]] = {
-    CatalogType.OBJECT: HatsCatalogLoader,
-    CatalogType.SOURCE: HatsCatalogLoader,
-    CatalogType.ASSOCIATION: AssociationCatalogLoader,
-    CatalogType.MARGIN: MarginCatalogLoader,
-}
+from lsdb.types import CatalogTypeVar
 
 
 def read_hats(
@@ -59,8 +60,6 @@ def read_hats(
         Catalog object loaded from the given parameters
     """
     # Creates a config object to store loading parameters from all keyword arguments.
-    # kwd_args = locals().copy()
-    # config_args = {field.name: kwd_args[field.name] for field in dataclasses.fields(HatsLoadingConfig)}
     config = HatsLoadingConfig(
         search_filter=search_filter,
         columns=columns,
@@ -69,10 +68,158 @@ def read_hats(
         kwargs=kwargs,
     )
 
-    catalog_info = TableProperties.read_from_dir(hc.io.file_io.get_upath(path))
-    catalog_type = catalog_info.catalog_type
+    hc_catalog = hc.read_hats(path)
+    if hc_catalog.schema is None:
+        raise ValueError(
+            "The catalog schema could not be loaded from metadata."
+            " Ensure your catalog has _common_metadata or _metadata files"
+        )
 
-    if catalog_type not in loader_class_for_catalog_type:
-        raise NotImplementedError(f"Cannot load catalog of type {catalog_type}")
-    loader_class = loader_class_for_catalog_type[catalog_type]
-    return loader_class(path, config).load_catalog()
+    catalog_type = hc_catalog.catalog_info.catalog_type
+
+    if catalog_type in (CatalogType.OBJECT, CatalogType.SOURCE):
+        return _load_object_catalog(hc_catalog, config)
+    if catalog_type == CatalogType.MARGIN:
+        return _load_margin_catalog(hc_catalog, config)
+    if catalog_type == CatalogType.ASSOCIATION:
+        return _load_assocation_catalog(hc_catalog, config)
+
+    raise NotImplementedError(f"Cannot load catalog of type {catalog_type}")
+
+
+def _load_assocation_catalog(hc_catalog, config):
+    """Load a catalog from the configuration specified when the loader was created
+
+    Returns:
+        Catalog object with data from the source given at loader initialization
+    """
+    if hc_catalog.catalog_info.contains_leaf_files:
+        dask_df, dask_df_pixel_map = _load_dask_df_and_map(hc_catalog, config)
+    else:
+        dask_meta_schema = _create_dask_meta_schema(hc_catalog.schema, config)
+        dask_df = nd.NestedFrame.from_pandas(dask_meta_schema, npartitions=1)
+        dask_df_pixel_map = {}
+    return AssociationCatalog(dask_df, dask_df_pixel_map, hc_catalog)
+
+
+def _load_margin_catalog(hc_catalog, config):
+    """Load a catalog from the configuration specified when the loader was created
+
+    Returns:
+        Catalog object with data from the source given at loader initialization
+    """
+    if config.search_filter:
+        filtered_catalog = config.search_filter.filter_hc_catalog(hc_catalog)
+        hc_catalog = hc.catalog.MarginCatalog(
+            filtered_catalog.catalog_info,
+            filtered_catalog.pixel_tree,
+            catalog_path=hc_catalog.catalog_path,
+            schema=filtered_catalog.schema,
+        )
+    dask_df, dask_df_pixel_map = _load_dask_df_and_map(hc_catalog, config)
+    margin = MarginCatalog(dask_df, dask_df_pixel_map, hc_catalog)
+    if config.search_filter is not None:
+        margin = margin.search(config.search_filter)
+    return margin
+
+
+def _load_object_catalog(hc_catalog, config):
+    """Load a catalog from the configuration specified when the loader was created
+
+    Returns:
+        Catalog object with data from the source given at loader initialization
+    """
+    if config.search_filter:
+        filtered_catalog = config.search_filter.filter_hc_catalog(hc_catalog)
+        if len(filtered_catalog.get_healpix_pixels()) == 0:
+            raise ValueError("The selected sky region has no coverage")
+        hc_catalog = hc.catalog.Catalog(
+            filtered_catalog.catalog_info,
+            filtered_catalog.pixel_tree,
+            catalog_path=hc_catalog.catalog_path,
+            moc=filtered_catalog.moc,
+            schema=filtered_catalog.schema,
+        )
+
+    dask_df, dask_df_pixel_map = _load_dask_df_and_map(hc_catalog, config)
+    catalog = Catalog(dask_df, dask_df_pixel_map, hc_catalog)
+    if config.search_filter is not None:
+        catalog = catalog.search(config.search_filter)
+    if isinstance(config.margin_cache, MarginCatalog):
+        catalog.margin = config.margin_cache
+        if config.search_filter is not None:
+            # pylint: disable=protected-access
+            catalog.margin = catalog.margin.search(config.search_filter)
+    elif config.margin_cache is not None:
+        hc_catalog = hc.read_hats(config.margin_cache)
+        catalog.margin = _load_margin_catalog(hc_catalog, config)
+    return catalog
+
+
+def _create_dask_meta_schema(schema: pa.Schema, config) -> npd.NestedFrame:
+    """Creates the Dask meta DataFrame from the HATS catalog schema."""
+    dask_meta_schema = schema.empty_table().to_pandas(types_mapper=config.get_dtype_mapper())
+    if (
+        dask_meta_schema.index.name != SPATIAL_INDEX_COLUMN
+        and SPATIAL_INDEX_COLUMN in dask_meta_schema.columns
+    ):
+        dask_meta_schema = dask_meta_schema.set_index(SPATIAL_INDEX_COLUMN)
+        if config.columns is not None and SPATIAL_INDEX_COLUMN in config.columns:
+            config.columns.remove(SPATIAL_INDEX_COLUMN)
+    if config.columns is not None:
+        dask_meta_schema = dask_meta_schema[config.columns]
+    return npd.NestedFrame(dask_meta_schema)
+
+
+def _load_dask_df_and_map(catalog: HCHealpixDataset, config) -> Tuple[nd.NestedFrame, DaskDFPixelMap]:
+    """Load Dask DF from parquet files and make dict of HEALPix pixel to partition index"""
+    pixels = catalog.get_healpix_pixels()
+    ordered_pixels = np.array(pixels)[get_pixel_argsort(pixels)]
+    divisions = get_pixels_divisions(ordered_pixels)
+    dask_meta_schema = _create_dask_meta_schema(catalog.schema, config)
+    if len(ordered_pixels) > 0:
+        ddf = nd.NestedFrame.from_map(
+            read_pixel,
+            ordered_pixels,
+            catalog=catalog,
+            query_url_params=config.make_query_url_params(),
+            columns=config.columns,
+            divisions=divisions,
+            meta=dask_meta_schema,
+            schema=catalog.schema,
+            **config.get_read_kwargs(),
+        )
+    else:
+        ddf = nd.NestedFrame.from_pandas(dask_meta_schema, npartitions=1)
+    pixel_to_index_map = {pixel: index for index, pixel in enumerate(ordered_pixels)}
+    return ddf, pixel_to_index_map
+
+
+def read_pixel(
+    pixel: HealpixPixel,
+    catalog: HCHealpixDataset,
+    *,
+    query_url_params: dict | None = None,
+    columns=None,
+    schema=None,
+    **kwargs,
+):
+    """Utility method to read a single pixel's parquet file from disk."""
+    if (
+        columns is not None
+        and schema is not None
+        and SPATIAL_INDEX_COLUMN in schema.names
+        and SPATIAL_INDEX_COLUMN not in columns
+    ):
+        columns = columns + [SPATIAL_INDEX_COLUMN]
+    dataframe = file_io.read_parquet_file_to_pandas(
+        hc.io.pixel_catalog_file(catalog.catalog_base_dir, pixel, query_url_params),
+        columns=columns,
+        schema=schema,
+        **kwargs,
+    )
+
+    if dataframe.index.name != SPATIAL_INDEX_COLUMN and SPATIAL_INDEX_COLUMN in dataframe.columns:
+        dataframe = dataframe.set_index(SPATIAL_INDEX_COLUMN)
+
+    return dataframe
